requirements.txt
flask~=1.1.2
dbus-python
zeroconf~=0.28.8
Flask-RESTful
numpy
opencv-python
pyautogui
pickledb
pyalsaaudio
psutil